# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "抽奖系统后端API",
    "description": "一个支持多种抽奖模式的完整抽奖系统API，包含认证、活动管理、参与者管理、抽奖功能、系统管理等模块。",
    "version": "1.0.0",
    "contact": {
      "name": "开发团队",
      "email": "buduan461@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "开发环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "webhookAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Webhook Token"
      }
    },
    "schemas": {
...
      "Prize": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "奖品ID"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "奖品名称"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "description": "奖品描述"
          },
          "total_quantity": {
            "type": "integer",
            "minimum": 0,
            "description": "奖品总数量"
          },
          "remaining_quantity": {
            "type": "integer",
            "minimum": 0,
            "description": "剩余数量"
          },
          "probability": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "中奖概率"
          },
          "sort_order": {
            "type": "integer",
            "minimum": 0,
            "description": "排序值"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          }
        },
        "required": ["success"]
      },
...
    },
    "responses": {
...
      "NotFoundError": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/lottery/activities/{id}": {
      "get": {
        "tags": ["抽奖模块"],
        "summary": "获取活动抽奖信息",
        "description": "获取活动抽奖信息（公开接口）",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "活动ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "activity": {
                              "$ref": "#/components/schemas/Activity"
                            },
                            "prizes": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Prize"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
    "/admin/activities": {
      "get": {
        "tags": ["管理员模块 - 活动管理"],
        "summary": "获取活动列表",
        "description": "获取当前用户创建的活动列表",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "搜索关键词",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "活动状态",
            "schema": {
              "type": "string",
              "enum": ["draft", "active", "ended"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "activities": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Activity"
                              }
                            },
                            "pagination": {
                              "$ref": "#/components/schemas/Pagination"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "post": {
        "tags": ["管理员模块 - 活动管理"],
        "summary": "创建新活动",
        "description": "创建新的抽奖活动",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "description": "活动名称"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "活动描述"
                  },
                  "lottery_mode": {
                    "type": "string",
                    "enum": ["offline", "online"],
                    "description": "抽奖模式"
                  },
                  "start_time": {
                    "type": "string",
                    "format": "date-time",
                    "description": "开始时间"
                  },
                  "end_time": {
                    "type": "string",
                    "format": "date-time",
                    "description": "结束时间"
                  },
                  "settings": {
                    "type": "object",
                    "properties": {
                      "max_lottery_codes": {
                        "type": "integer",
                        "minimum": 1,
                        "description": "最大抽奖码数量"
                      },
                      "lottery_code_format": {
                        "type": "string",
                        "enum": ["4_digit_number", "8_digit_number", "8_digit_alphanumeric", "12_digit_number", "12_digit_alphanumeric"],
                        "description": "抽奖码格式"
                      },
                      "allow_duplicate_phone": {
                        "type": "boolean",
                        "description": "是否允许重复手机号"
                      }
                    }
                  }
                },
                "required": ["name", "lottery_mode"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "活动创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "activity": {
                              "$ref": "#/components/schemas/Activity"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/admin/activities/{id}": {
      "get": {
        "tags": ["管理员模块 - 活动管理"],
        "summary": "获取活动详情",
        "description": "获取指定活动的详细信息",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "活动ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "activity": {
                              "allOf": [
                                {
                                  "$ref": "#/components/schemas/Activity"
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "lottery_codes_count": {
                                      "type": "integer",
                                      "description": "抽奖码总数"
                                    },
                                    "prizes": {
                                      "type": "array",
                                      "items": {
                                        "$ref": "#/components/schemas/Prize"
                                      }
                                    }
                                  }
                                }
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "put": {
        "tags": ["管理员模块 - 活动管理"],
        "summary": "更新活动信息",
        "description": "更新指定活动的信息",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "活动ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "description": "活动名称"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "活动描述"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "active", "ended"],
                    "description": "活动状态"
                  },
                  "start_time": {
                    "type": "string",
                    "format": "date-time",
                    "description": "开始时间"
                  },
                  "end_time": {
                    "type": "string",
                    "format": "date-time",
                    "description": "结束时间"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "activity": {
                              "$ref": "#/components/schemas/Activity"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "delete": {
        "tags": ["管理员模块 - 活动管理"],
        "summary": "删除活动",
        "description": "删除指定活动（仅限草稿状态）",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "活动ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
    "/admin/activities/{id}/lottery-codes/batch": {
      "post": {
        "tags": ["管理员模块 - 活动管理"],
        "summary": "批量创建抽奖码",
        "description": "为活动批量创建抽奖码",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "活动ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "count": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000,
                    "description": "创建数量"
                  }
                },
                "required": ["count"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "批量创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "created_count": {
                              "type": "integer",
                              "description": "成功创建的数量"
                            },
                            "lottery_codes": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/LotteryCode"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  },
...
}